+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v0.17.0] - ???
+
+### Added
+
+- Data exporter can parse composite question titles or option labels (when text contains multiple parts).
+- Data exporter logic to handle question type: "responsive single choice array"
+
 ## [v0.16.2] - 2021-06-16
 
 ### Changed